13872: Improve handling of address bar display for IPFS and IPNS
We should change this to only keep ipfs:// and ipns:// when the address being loaded when:
i) Using the Brave node and the port matches the Brave node's port.
ii) Using the public node and the port matches the public node, and it is localhost.
iii) When user enters custom public gateway Brave should always replace raw IP (127.0.0.1) with localhost,
as that enables subdomains and origin isolation. Raw IP can't have subdomains and
all content roots share the same origin, which we don't want in browser.